+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 ## [Unreleased]
 ### Changed
 - `Stream::consume_reference` will return only `InvalidReference` error from now.
+- `Error::InvalidTokenWithCause` merged into `Error::InvalidToken`.
 
 ### Removed
 - `failure` dependency.